+++ services/auth-service/internal/infrastructure/persistence/postgres/model/user_role_po.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type UserRole struct {
-	UserID    uint
-	RoleID    uint
+	UserID    int64
+	RoleID    int64
 	GrantedAt time.Time
 }
