+++ services/user-service/internal/domain/profile/repository.go
@@ -3,5 +3,5 @@ package profile
 import "context"
 
 type Repository interface {
-	FindByUserID(ctx context.Context, userID uint) (*Profile, error)
+	FindByUserID(ctx context.Context, userID int64) (*Profile, error)
 }
