+++ services/user-service/internal/infrastructure/persistence/postgres/repository/profile_repo.go
@@ -19,7 +19,7 @@ func NewProfileRepo(db *gorm.DB) profile.Repository {
 	return &ProfileRepo{db: db}
 }
 
-func (r *ProfileRepo) FindByUserID(ctx context.Context, userID uint) (*profile.Profile, error) {
+func (r *ProfileRepo) FindByUserID(ctx context.Context, userID int64) (*profile.Profile, error) {
 	if userID == 0 {
 		return nil, profile.ErrInvalidUserID
 	}
@@ -32,7 +32,7 @@ func (r *ProfileRepo) FindByUserID(ctx context.Context, userID uint) (*profile.P
 		return nil, fmt.Errorf("failed to query user %d's profile %w", userID, err)
 	}
 
-	addIDs := make([]uint, 0)
+	addIDs := make([]int64, 0)
 	err = r.db.WithContext(ctx).Model(&model.UserAddress{}).Where("user_id = ?", userID).Pluck("id", &addIDs).Error
 	if err != nil {
 		return nil, fmt.Errorf("failed to query user %d's profile %w", userID, err)
